moved views to proper package
