+++ brav/app/src/main/AndroidManifest.xml
@@ -11,9 +11,9 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name=".MediatorDetailActivity">
+        <activity android:name=".views.MediatorDetailActivity">
         </activity>
-        <activity android:name=".CaseDetailsActivity">
+        <activity android:name=".views.CaseDetailsActivity">
         </activity>
         <activity android:name=".views.UserAccountActivity">
         </activity>
