+++ brav/app/src/main/java/com/thadocizn/brav/adapters/CaseAdapter.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.thadocizn.brav.CaseDetailsActivity
+import com.thadocizn.brav.views.CaseDetailsActivity
 import com.thadocizn.brav.R
 import com.thadocizn.brav.models.Case
 import kotlinx.android.synthetic.main.list_item_case.view.*
