+++ brav/app/src/main/java/com/thadocizn/brav/adapters/MediatorAdapter.kt
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import android.widget.Button
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.thadocizn.brav.MediatorDetailActivity
+import com.thadocizn.brav.views.MediatorDetailActivity
 import com.thadocizn.brav.views.UserAccountActivity
 import com.thadocizn.brav.R
 import com.thadocizn.brav.models.Mediator
