+++ brav/app/src/main/res/layout/activity_case_details.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".CaseDetailsActivity" android:orientation="vertical">
+        tools:context=".views.CaseDetailsActivity" android:orientation="vertical">
 
     <LinearLayout
             android:orientation="horizontal"
