+++ brav/app/src/main/res/layout/activity_mediator_detail.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MediatorDetailActivity" android:orientation="vertical">
+        tools:context=".views.MediatorDetailActivity" android:orientation="vertical">
 
     <LinearLayout
             android:orientation="horizontal"
